Rename storage_proof module
