+++ src/network/protocol.rs
@@ -33,7 +33,7 @@ mod grandpa;
 mod grandpa_warp_sync;
 mod identify;
 mod state_request;
-mod storage_proof;
+mod storage_call_proof;
 
 pub use self::block_announces::*;
 pub use self::block_request::*;
@@ -41,4 +41,4 @@ pub use self::grandpa::*;
 pub use self::grandpa_warp_sync::*;
 pub use self::identify::*;
 pub use self::state_request::*;
-pub use self::storage_proof::*;
+pub use self::storage_call_proof::*;
